Move useDebounce in hooks folder
